Bugfix: Various runtime errors from goToPose (#46)
* Fixed issue with storing baudrate in an undersized integer

* Added the missing go_to_pose instantiation in motionControl

* Added missing reset for rotational PID when other PIDs are reset

+++ Driftless_PushBack_PROS/include/driftless/config/BlueConfig.hpp
@@ -212,7 +212,7 @@ class BlueConfig : public IConfig {
 
   // ## COPROCESSOR ##
 
-  static constexpr uint16_t COPROCESSOR_BAUD_RATE{115200};
+  static constexpr int32_t COPROCESSOR_BAUD_RATE{115200};
 
   // ## DRIVE ##
